sqlite: implement size in .get_info method
The type of the size is changed from int32_t to int64_t.
An int32_t can only hold sizes up to 2GB which is not
enough for a database. A int64_t is more than enough to
hold a size.

+++ rbh-info/rbh-info.c
@@ -275,7 +275,7 @@ _get_collection_avg_obj_size(const struct rbh_value *value)
 static void
 _get_collection_size(const struct rbh_value *value)
 {
-    printf("%d\n", value->int32);
+    printf("%ld\n", value->int64);
 }
 
 static void
